# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PPIS API DOCUMENTATION",
        "description": "This is the Project Progres Information System API documentation.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/assets": {
            "get": {
                "tags": [
                    "asset"
                ],
                "summary": "Get a list of assets",
                "operationId": "62a8049651a178012b0a8d9a7c66fc7e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "asset_id": {
                                                        "type": "integer"
                                                    },
                                                    "asset_name": {
                                                        "type": "string"
                                                    },
                                                    "uuid": {
                                                        "type": "string"
                                                    },
                                                    "serial_number": {
                                                        "type": "string"
                                                    },
                                                    "code": {
                                                        "type": "string"
                                                    },
                                                    "created_by": {
                                                        "type": "integer"
                                                    },
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "middle_name": {
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "asset"
                ],
                "summary": "Store a new assets",
                "operationId": "e563897b93239b1f28030ce31eda95d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "asset_name": {
                                        "type": "string"
                                    },
                                    "serial_number": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}